Add IndexController update sendLatestNoticias
