+++ src/routes/index.js
@@ -1,11 +1,13 @@
 const express = require("express");
 const router = express.Router();
 const NoticiaController = require("../controllers/NoticiaController");
+const IndexController = require("../controllers/IndexController");
 const nodemailer = require("nodemailer");
-const GMAIL_USER = process.env.GMAIL_USER;
-const GMAIL_PASS = process.env.GMAIL_PASS;
 
-router.get("/", NoticiaController.indexLatestNoticias);
+router.get("/", async (req, res) => {
+  const noticias = await NoticiaController.sendLatestNoticias();
+  res.render("index", { noticias });
+});
 
 router.get("/faqs", (req, res) => {
   res.render("faqs");
@@ -16,33 +18,12 @@ router.get("/contacto", (req, res) => {
 });
 
 // POST del form contacto
-router.post("/contacto", (req, res) => {
-  // MTP server
-  const smtpTrans = nodemailer.createTransport({
-    host: "smtp.gmail.com",
-    port: 465,
-    secure: true,
-    auth: {
-      user: GMAIL_USER,
-      pass: GMAIL_PASS,
-    },
-  });
+router.post("/contacto", IndexController.sendContactUsEmail);
 
-  // Correo que se mandará
-  const mail = {
-    from: "Your sender info here", // This is ignored by Gmail
-    to: GMAIL_USER,
-    subject: "Nuevo mensaje de contacto de ISND",
-    text: `${req.body.nombre} (${req.body.correo}) says: ${req.body.mensaje}`,
-  };
-
-  // Enviar correo
-  smtpTrans.sendMail(mail, (error, response) => {
-    error ? res.render("contacto", { error: true }) : res.redirect("/");
-  });
+router.get("/nosotros", async (req, res) => {
+  const noticias = NoticiaController.sendLatestNoticias();
+  res.render("nosotros", { noticias });
 });
-
-router.get("/nosotros", NoticiaController.nosotrosLatestNoticias);
 
 router.get("/proyectos", (req, res) => {
   res.render("proyectos");
